+++ src/main/java/mirsario/cameraoverhaul/abstractions/TextAbstractions.java
@@ -8,18 +8,13 @@ package mirsario.cameraoverhaul.abstractions;
 import net.minecraft.network.chat.*;
 
 public final class TextAbstractions {
-	public static Component createText(String key) {
+	public static Component getText(String key) {
 		//? if >=1.19 {
 			return Component.translatable(key);
 		//?} else
 			/*return new TranslatableComponent(key);*/
 	}
 
-	public static String getTextValue(String key) { return TextAbstractions.createText(key).getString(); }
-
-	//? if >=1.17 {
-	public static net.minecraft.network.chat.Component getText(String key) { return TextAbstractions.createText(key); }
-	//?} else
-	/*public static String getText(String key) { return TextAbstractions.createText(key).getString(); }*/
+	public static String getTextValue(String key) { return getText(key).getString(); }
 }
 //?}